+++ src/solver/type_getters.jl
@@ -136,7 +136,6 @@ function get_spaces(parsed_args, params, comms_ctx)
     z_elem = Int(parsed_args["z_elem"])
     z_max = FT(parsed_args["z_max"])
     dz_bottom = FT(parsed_args["dz_bottom"])
-    dz_top = FT(parsed_args["dz_top"])
     topography = parsed_args["topography"]
     bubble = parsed_args["bubble"]
     deep = parsed_args["deep_atmosphere"]
@@ -184,7 +183,7 @@ function get_spaces(parsed_args, params, comms_ctx)
         h_space =
             make_horizontal_space(horizontal_mesh, quad, comms_ctx, bubble)
         z_stretch = if parsed_args["z_stretch"]
-            Meshes.GeneralizedExponentialStretching(dz_bottom, dz_top)
+            Meshes.HyperbolicTangentStretching(dz_bottom)
         else
             Meshes.Uniform()
         end
@@ -219,7 +218,7 @@ function get_spaces(parsed_args, params, comms_ctx)
         h_space =
             make_horizontal_space(horizontal_mesh, quad, comms_ctx, bubble)
         z_stretch = if parsed_args["z_stretch"]
-            Meshes.GeneralizedExponentialStretching(dz_bottom, dz_top)
+            Meshes.HyperbolicTangentStretching(dz_bottom)
         else
             Meshes.Uniform()
         end
@@ -241,7 +240,7 @@ function get_spaces(parsed_args, params, comms_ctx)
         h_space =
             make_horizontal_space(horizontal_mesh, quad, comms_ctx, bubble)
         z_stretch = if parsed_args["z_stretch"]
-            Meshes.GeneralizedExponentialStretching(dz_bottom, dz_top)
+            Meshes.HyperbolicTangentStretching(dz_bottom)
         else
             Meshes.Uniform()
         end
@@ -265,7 +264,7 @@ function get_spaces(parsed_args, params, comms_ctx)
         h_space =
             make_horizontal_space(horizontal_mesh, quad, comms_ctx, bubble)
         z_stretch = if parsed_args["z_stretch"]
-            Meshes.GeneralizedExponentialStretching(dz_bottom, dz_top)
+            Meshes.HyperbolicTangentStretching(dz_bottom)
         else
             Meshes.Uniform()
         end